# static/swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "Student Meal Express",
      "version": "1.0.0"
    },
    "paths": {
      "/users/all": {
        "get": {
          "description": "Gets all users",
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "Successful operation"
            }
          }
        }
      },
      "/users": {
        "get": {
          "description": "Gets user by id",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "user_id",
              "in": "query",
              "description": "JSON data for the api",
              "required": true,
              "schema": {
                  "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful operation"
            }
          }
        },
        "post": {
          "description": "Creates new user",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "user data",
              "in": "body",
              "description": "JSON data for the api",
              "required": true,
              "schema": {
                "$ref": "#/definitions/UserData"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful operation"
            },
            "400": {
              "description": "Invalid request data"
            }
          }
        },
        "put": {
          "description": "Updates a user",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "user data",
              "in": "body",
              "description": "JSON data for the api",
              "required": true,
              "schema": {
                "$ref": "#/definitions/UserData"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful operation"
            },
            "400": {
              "description": "Invalid request data"
            }
          }
        },
        "delete": {
          "description": "Deletes a user",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "user_id",
              "in": "query",
              "description": "JSON data for the api",
              "required": true,
              "schema": {
                  "type": "string"
              }
            }
          ]
        }
      },
      "/recipes/all": {
        "get": {
          "description": "Gets all recipes",
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "Successful operation"
            }
          }
        }
      },
      "/recipes": {
        "get": {
          "description": "Gets recipe by id",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "recipe_id",
              "in": "query",
              "description": "JSON data for the api",
              "required": true,
              "schema": {
                  "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful operation"
            }
          }
        },
        "post": {
          "description": "Creates new recipe",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "recipe data",
              "in": "body",
              "description": "JSON data for the api",
              "required": true,
              "schema": {
                "$ref": "#/definitions/RecipeData"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful operation"
            },
            "400": {
              "description": "Invalid request data"
            }
          }
        },
        "put": {
          "description": "Updates a recipe",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "recipe data",
              "in": "body",
              "description": "JSON data for the api",
              "required": true,
              "schema": {
                "$ref": "#/definitions/RecipeData"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful operation"
            },
            "400": {
              "description": "Invalid request data"
            }
          }
        },
        "delete": {
          "description": "Deletes a recipe",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "recipe_id",
              "in": "query",
              "description": "JSON data for the api",
              "required": true,
              "schema": {
                  "type": "string"
              }
            }
          ]
        }
      },
      "/tags/all": {
        "get": {
          "description": "Gets all tags",
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "Successful operation"
            }
          }
        }
      },
      "/tags": {
        "get": {
          "description": "Gets tag by id",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "tag_id",
              "in": "query",
              "description": "JSON data for the api",
              "required": true,
              "schema": {
                  "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful operation"
            }
          }
        },
        "post": {
          "description": "Creates new tag",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "tag data",
              "in": "body",
              "description": "JSON data for the api",
              "required": true,
              "schema": {
                "$ref": "#/definitions/TagData"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful operation"
            },
            "400": {
              "description": "Invalid request data"
            }
          }
        },
        "put": {
          "description": "Updates a tag",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "tag data",
              "in": "body",
              "description": "JSON data for the api",
              "required": true,
              "schema": {
                "$ref": "#/definitions/TagData"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful operation"
            },
            "400": {
              "description": "Invalid request data"
            }
          }
        },
        "delete": {
          "description": "Deletes a tag",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "tag_id",
              "in": "query",
              "description": "JSON data for the api",
              "required": true,
              "schema": {
                  "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful operation"
            },
            "400": {
              "description": "Invalid request data"
            }
          }
        }
      }
    },
    "definitions": {
      "UserData": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "integer"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password_hash": {
            "type": "string"
          },
          "join_date": {
            "type": "string"
          },
          "is_active": {
            "type": "boolean"
          }
        }
      },
      "RecipeData": {
        "type": "object",
        "properties": {
          "recipe_id": {
            "type": "integer"
          },
          "cook_time": {
            "type": "integer"
          },
          "created_by": {
            "type": "integer"
          },
          "creation_date": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "serving_size": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "TagData": {
        "type": "object",
        "properties": {
          "tag_id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      }
    }
  }